Added priject "Rectangle" to the Lab#11

+++ Lab11/RPN-Stack/RPN-Stack/GenericStack.cpp
@@ -57,7 +57,7 @@ T GenericStack<T>::readTop() {
 }
 
 template <typename T>
-int GenericStack<T>::getSize(){
+unsigned long GenericStack<T>::getSize(){
     return stack.size();
 }
